add run feature flag to fantoch
