+++ fantoch/src/lib.rs
@@ -39,7 +39,38 @@ pub mod sim;
 
 // This module contains the definition of Runner` (that actually runs a given
 // `Process`)
+#[cfg(feature = "run")]
 pub mod run;
+
+pub mod load_balance {
+    use crate::id::Dot;
+
+    // the worker index that should be used by leader-based protocols
+    pub const LEADER_WORKER_INDEX: usize = 0;
+
+    // the worker index that should be for garbage collection:
+    // - it's okay to be the same as the leader index because this value is not
+    //   used by leader-based protocols
+    // - e.g. in fpaxos, the gc only runs in the acceptor worker
+    pub const GC_WORKER_INDEX: usize = 0;
+
+    pub const WORKERS_INDEXES_RESERVED: usize = 2;
+
+    pub fn worker_index_no_shift(index: usize) -> Option<(usize, usize)> {
+        // when there's no shift, the index must be either 0 or 1
+        assert!(index < WORKERS_INDEXES_RESERVED);
+        Some((0, index))
+    }
+
+    // note: reserved indexing always reserve the first two workers
+    pub const fn worker_index_shift(index: usize) -> Option<(usize, usize)> {
+        Some((WORKERS_INDEXES_RESERVED, index))
+    }
+
+    pub fn worker_dot_index_shift(dot: &Dot) -> Option<(usize, usize)> {
+        worker_index_shift(dot.sequence() as usize)
+    }
+}
 
 // This module contains some utilitary functions.
 pub mod util;
