+++ fantoch_ps/Cargo.toml
@@ -23,7 +23,7 @@ rand = "0.8.0"
 rayon = { version = "1.5.0", optional = true }
 serde = { version = "1.0.118", features = ["derive"] }
 threshold = "0.8.15"
-tokio = { version = "1.0.0", features = ["full", "parking_lot"] }
+tokio = { version = "1.0.1", features = ["full", "parking_lot"] }
 tracing = "0.1.22"
 tracing-appender = "0.1.1"
 parking_lot = "0.11.1"
